requirements.txt
torch==2.3.1
torchvision==0.18.1
matplotlib==3.7.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
Pillow==10.0.0

